+++ CHANGELOG.md
@@ -3,8 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
-### Fixed
-- Ensure `Model::definition()` is always called in static context
+### Added
+- The new `ListenerException` class can be thrown from event listeners as a shortcut to stop events.
 
 ### Removed
 - `Model::getDispatcher()` was removed and replaced with the `EventManager` class
@@ -47,6 +47,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - Use a strict equality check when casting an empty string to null on a nullable property. Previously this would check for a falsey value.
+- Ensure `Model::definition()` is always called in static context
 
 ### Removed
 - Removed `Errors::setLocale`, `Errors::getLocale()`, and `Errors::setGlobalLocale()`. It is now required to use `Errors::setTranslator()`.